fix: clean up and some minor fixes
