+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Platforms/iOS/BiometricService.ios.cs
