+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Shared/AuthenticationResponse.cs
@@ -3,8 +3,6 @@
 public class AuthenticationResponse
 {
     public BiometricResponseStatus Status { get; set; }
-
     public AuthenticationType AuthenticationType { get; set; }
-
     public string ErrorMsg { get; set; }
 }
