+++ Plugin.Maui.Biometric/Plugin.Maui.Biometric/Shared/IBiometric.cs
@@ -3,6 +3,5 @@
 public interface IBiometric
 {
     Task<BiometricHwStatus> GetAuthenticationStatusAsync(AuthenticatorStrength authStrength = AuthenticatorStrength.Strong);
-
     Task<AuthenticationResponse> AuthenticateAsync(AuthenticationRequest request, CancellationToken token);
 }
